CH2/2-6/pinecone_intro/pyproject.toml
[tool.poetry]
name = "pinecone-intro"
version = "0.1.0"
description = ""
authors = ["KoKo Mexcelsa <qqkerk@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.8.8"
pinecone-client = "^3.0.2"
openai = "^1.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
